consolidate duplicated functions to OrganizationDashboardWidgetTestCase
