+++ tests/sentry/api/endpoints/test_organization_dashboard_details.py
@@ -58,18 +58,6 @@ class OrganizationDashboardDetailsTestCase(OrganizationDashboardWidgetTestCase):
             kwargs={"organization_slug": self.organization.slug, "dashboard_id": dashboard_id},
         )
 
-    def assert_serialized_widget(self, data, expected_widget):
-        if "id" in data:
-            assert data["id"] == six.text_type(expected_widget.id)
-        if "title" in data:
-            assert data["title"] == expected_widget.title
-        if "interval" in data:
-            assert data["interval"] == expected_widget.interval
-        if "displayType" in data:
-            assert data["displayType"] == DashboardWidgetDisplayTypes.get_type_name(
-                expected_widget.display_type
-            )
-
     def assert_serialized_dashboard(self, data, dashboard):
         assert data["id"] == six.text_type(dashboard.id)
         assert data["title"] == dashboard.title
@@ -136,9 +124,6 @@ class OrganizationDashboardDetailsPutTest(OrganizationDashboardDetailsTestCase):
             display_type=DashboardWidgetDisplayTypes.LINE_CHART,
         )
         self.widget_ids = [self.widget_1.id, self.widget_2.id, self.widget_3.id, self.widget_4.id]
-
-    def get_widgets(self, dashboard_id):
-        return DashboardWidget.objects.filter(dashboard_id=dashboard_id).order_by("order")
 
     def get_widget_queries(self, widget):
         return DashboardWidgetQuery.objects.filter(widget=widget).order_by("order")
